requirements.txt
rx
requests
channels-redis
channels
python-dotenv